requirements.txt
Django==2.2.5
psycopg2==2.8.3
django-heroku==0.3.1
gunicorn==19.9.0
Pillow==6.1.0
whitenoise==4.1.3
python-decouple==3.1
dj-database-url==0.5.0
requests==2.22.0
python-dateutil==2.8.1
